refactor to RwCustomQueryPlanner
